APIGOV-22764_tj update
