+++ pkg/agent/poller/client.go
@@ -8,7 +8,6 @@ import (
 	"github.com/Axway/agent-sdk/pkg/agent/handler"
 	"github.com/Axway/agent-sdk/pkg/config"
 	"github.com/Axway/agent-sdk/pkg/harvester"
-	"github.com/Axway/agent-sdk/pkg/util"
 	"github.com/Axway/agent-sdk/pkg/util/errors"
 	hc "github.com/Axway/agent-sdk/pkg/util/healthcheck"
 	"github.com/Axway/agent-sdk/pkg/watchmanager/proto"
@@ -35,33 +34,6 @@ type harvesterConfig struct {
 }
 
 type onClientStopCb func()
-
-type ClientOpt func(pc *PollClient)
-
-// WithHarvester configures the polling client to use harvester
-func WithHarvester(hClient harvester.Harvest, sequence events.SequenceProvider, topicSelfLink string) ClientOpt {
-	return func(pc *PollClient) {
-		pc.harvesterConfig.hClient = hClient
-		pc.harvesterConfig.topicSelfLink = topicSelfLink
-		pc.harvesterConfig.sequence = sequence
-	}
-}
-
-// WithOnClientStop func to execute when the client shuts down
-func WithOnClientStop(cb func()) ClientOpt {
-	return func(pc *PollClient) {
-		pc.onClientStop = cb
-	}
-}
-
-// WithOnConnect func to execute when a connection to central is made
-func WithOnConnect() ClientOpt {
-	return func(pc *PollClient) {
-		pc.onStreamConnection = func() {
-			hc.RegisterHealthcheck(util.AmplifyCentral, "central", pc.Healthcheck)
-		}
-	}
-}
 
 // NewPollClient creates a polling client
 func NewPollClient(
