+++ pkg/agent/stream/options.go
@@ -1,0 +1,51 @@
+package stream
+
+import (
+	agentcache "github.com/Axway/agent-sdk/pkg/agent/cache"
+	"github.com/Axway/agent-sdk/pkg/agent/events"
+	"github.com/Axway/agent-sdk/pkg/apic/apiserver/models/management/v1alpha1"
+	"github.com/Axway/agent-sdk/pkg/harvester"
+	"github.com/Axway/agent-sdk/pkg/util"
+	hc "github.com/Axway/agent-sdk/pkg/util/healthcheck"
+)
+
+// StreamOpt func for setting fields on the StreamerClient
+type StreamOpt func(client *StreamerClient)
+
+// WithWatchTopic sets the watch topic
+func WithWatchTopic(wt *v1alpha1.WatchTopic) StreamOpt {
+	return func(client *StreamerClient) {
+		client.wt = wt
+		client.topicSelfLink = wt.GetSelfLink()
+	}
+}
+
+func WithCacheManager(cache agentcache.Manager) StreamOpt {
+	return func(client *StreamerClient) {
+		client.cacheManager = cache
+	}
+}
+
+// WithHarvester configures the streaming client to use harvester for syncing initial events
+func WithHarvester(hClient harvester.Harvest, sequence events.SequenceProvider) StreamOpt {
+	return func(client *StreamerClient) {
+		client.sequence = sequence
+		client.harvester = hClient
+	}
+}
+
+// WithEventSyncError sets the callback func to run when there is an error syncing events
+func WithEventSyncError(cb func()) StreamOpt {
+	return func(client *StreamerClient) {
+		client.onEventSyncError = cb
+	}
+}
+
+// WithOnStreamConnection func to execute when a connection to central is made
+func WithOnStreamConnection() StreamOpt {
+	return func(pc *StreamerClient) {
+		pc.onStreamConnection = func() {
+			hc.RegisterHealthcheck(util.AmplifyCentral, "central", pc.Healthcheck)
+		}
+	}
+}
